move fixtures to the right folder
